rds_client.job has enclave functionality
