+++ src/syft_rds/client/rds_clients/job.py
@@ -7,6 +7,7 @@ from uuid import UUID
 
 from loguru import logger
 
+from syft_core import Client
 from syft_rds.client.exceptions import RDSValidationError
 from syft_rds.client.rds_clients.base import RDSClientModule
 from syft_rds.client.utils import PathLike
@@ -37,6 +38,7 @@ class JobRDSClient(RDSClientModule[Job]):
         runtime_name: str | None = None,
         runtime_kind: str | None = None,
         runtime_config: dict | None = None,
+        enclave: str = "",
     ) -> Job:
         """`submit` is a convenience method to create both a UserCode and a Job in one call."""
         if custom_function is not None:
@@ -70,6 +72,7 @@ class JobRDSClient(RDSClientModule[Job]):
             tags=tags,
             custom_function=custom_function,
             runtime=runtime,
+            enclave=enclave,
         )
 
         return job
@@ -143,6 +146,17 @@ class JobRDSClient(RDSClientModule[Job]):
                 f"Invalid user_code type {type(user_code)}. Must be UserCode, UUID, or str"
             )
 
+    def _verify_enclave(self, enclave: str) -> None:
+        """Verify that the enclave is valid."""
+        client: Client = self.rpc.connection.sender_client
+        enclave_app_dir = client.app_data("enclave", datasite=enclave)
+        public_key_path = enclave_app_dir / "keys" / "public_key.pem"
+        if not public_key_path.exists():
+            raise RDSValidationError(
+                f"Enclave {enclave} does not exist or is not valid. "
+                f"Public key file {public_key_path} not found."
+            )
+
     def create(
         self,
         user_code: UserCode | UUID,
@@ -152,10 +166,13 @@ class JobRDSClient(RDSClientModule[Job]):
         description: str | None = None,
         tags: list[str] | None = None,
         custom_function: CustomFunction | UUID | None = None,
+        enclave: str = "",
     ) -> Job:
-        # TODO ref dataset by UID instead of name
         user_code_id = self._resolve_usercode_id(user_code)
         custom_function_id = self._resolve_custom_func_id(custom_function)
+
+        if enclave:
+            self._verify_enclave(enclave)
 
         job_create = JobCreate(
             name=name,
@@ -165,6 +182,7 @@ class JobRDSClient(RDSClientModule[Job]):
             runtime_id=runtime.uid,
             dataset_name=dataset_name,
             custom_function_id=custom_function_id,
+            enclave=enclave,
         )
         job = self.rpc.job.create(job_create)
 
@@ -252,6 +270,14 @@ class JobRDSClient(RDSClientModule[Job]):
                 f"Job {job.uid} is not shared. Current status: {job.status}"
             )
         return self._get_results_from_dir(job, job.output_path)
+
+    def approve(self, job: Job) -> Job:
+        if not self.is_admin:
+            raise RDSValidationError("Only admins can approve jobs")
+        job_update = job.get_update_for_approve()
+        updated_job = self.rpc.job.update(job_update)
+        job.apply_update(updated_job, in_place=True)
+        return job
 
     def reject(self, job: Job, reason: str = "Unspecified") -> None:
         if not self.is_admin:
